// angular-frontend/src/app/models/positions.model.ts
export default class Positions {
    id?: number | null;
    positions?: string | null;
    lsl?: number | null;
    Ll?: number | null;
    ast?: number | null;
    At?: number | null;
    bst?: number | null;
    Bt?: number | null;
    cst?: number | null;
    Ct?: number | null;
    dst?: number | null;
    Dt?: number | null;
    est?: number | null;
    Et?: number | null;
    fst?: number | null;
    Ft?: number | null;
    lsr?: number | null;
    Lr?: number | null;
    asb?: number | null;
    Ab?: number | null;
    bsb?: number | null;
    Bb?: number | null;
    csb?: number | null;
    Cb?: number | null;
    dsb?: number | null;
    Db?: number | null;
    esb?: number | null;
    Eb?: number | null;
    fsb?: number | null;
    Fb?: number | null;
}